Begin adding backend support for record ids when distinct
