+++ specifyweb/stored_queries/execution.py
@@ -18,104 +18,16 @@ from .format import ObjectFormatter
 from .query_construct import QueryConstruct
 from .queryfield import QueryField
 from .relative_date_utils import apply_absolute_date
-from .field_spec_maps import apply_specify_user_name
+from .utils import apply_specify_user_name, set_group_concat_max_len, filter_by_collection, can_uniquely_identify
 from ..notifications.models import Message
 from ..permissions.permissions import check_table_permissions
 from ..specify.auditlog import auditlog
+from ..specify.datamodel import datamodel
 from ..specify.models import Loan, Loanpreparation, Loanreturnpreparation
 
 logger = logging.getLogger(__name__)
 
 SORT_TYPES = [None, asc, desc]
-
-def set_group_concat_max_len(session):
-    """The default limit on MySQL group concat function is quite
-    small. This function increases it for the database connection for
-    the given session.
-    """
-    session.connection().execute('SET group_concat_max_len = 1024 * 1024 * 1024')
-
-def filter_by_collection(model, query, collection):
-    """Add predicates to the given query to filter result to items scoped
-    to the given collection. The model argument indicates the "base"
-    table of the query. E.g. If model was CollectingEvent, this
-    function would limit the results to collecting events in the same
-    discipline as the given collection since collecting events are
-    scoped to the discipline level.
-    """
-    if (model is models.Accession and
-        collection.discipline.division.institution.isaccessionsglobal):
-        logger.info("not filtering query b/c accessions are global in this database")
-        return query
-
-    if model is models.Taxon:
-        logger.info("filtering taxon to discipline: %s", collection.discipline.name)
-        return query.filter(model.TaxonTreeDefID == collection.discipline.taxontreedef_id)
-
-    if model is models.TaxonTreeDefItem:
-        logger.info("filtering taxon rank to discipline: %s", collection.discipline.name)
-        return query.filter(model.TaxonTreeDefID == collection.discipline.taxontreedef_id)
-
-    if model is models.Geography:
-        logger.info("filtering geography to discipline: %s", collection.discipline.name)
-        return query.filter(model.GeographyTreeDefID == collection.discipline.geographytreedef_id)
-
-    if model is models.GeographyTreeDefItem:
-        logger.info("filtering geography rank to discipline: %s", collection.discipline.name)
-        return query.filter(model.GeographyTreeDefID == collection.discipline.geographytreedef_id)
-
-    if model is models.LithoStrat:
-        logger.info("filtering lithostrat to discipline: %s", collection.discipline.name)
-        return query.filter(model.LithoStratTreeDefID == collection.discipline.lithostrattreedef_id)
-
-    if model is models.LithoStratTreeDefItem:
-        logger.info("filtering lithostrat rank to discipline: %s", collection.discipline.name)
-        return query.filter(model.LithoStratTreeDefID == collection.discipline.lithostrattreedef_id)
-
-    if model is models.GeologicTimePeriod:
-        logger.info("filtering geologic time period to discipline: %s", collection.discipline.name)
-        return query.filter(model.GeologicTimePeriodTreeDefID == collection.discipline.geologictimeperiodtreedef_id)
-
-    if model is models.GeologicTimePeriodTreeDefItem:
-        logger.info("filtering geologic time period rank to discipline: %s", collection.discipline.name)
-        return query.filter(model.GeologicTimePeriodTreeDefID == collection.discipline.geologictimeperiodtreedef_id)
-
-    if model is models.Storage:
-        logger.info("filtering storage to institution: %s", collection.discipline.division.institution.name)
-        return query.filter(model.StorageTreeDefID == collection.discipline.division.institution.storagetreedef_id)
-
-    if model is models.StorageTreeDefItem:
-        logger.info("filtering storage rank to institution: %s", collection.discipline.division.institution.name)
-        return query.filter(model.StorageTreeDefID == collection.discipline.division.institution.storagetreedef_id)
-
-    if model in (
-            models.Agent,
-            models.Accession,
-            models.RepositoryAgreement,
-            models.ExchangeIn,
-            models.ExchangeOut,
-            models.ConservDescription,
-    ):
-        return query.filter(model.DivisionID == collection.discipline.division_id)
-
-    for filter_col, scope, scope_name in (
-            ('CollectionID'       , lambda collection: collection, lambda o: o.collectionname),
-            ('collectionMemberId' , lambda collection: collection, lambda o: o.collectionname),
-            ('DisciplineID'       , lambda collection: collection.discipline, lambda o: o.name),
-
-        # The below are disabled to match Specify 6 behavior.
-            # ('DivisionID'         , lambda collection: collection.discipline.division, lambda o: o.name),
-            # ('InstitutionID'      , lambda collection: collection.discipline.division.institution, lambda o: o.name),
-    ):
-
-        if hasattr(model, filter_col):
-            o = scope(collection)
-            logger.info("filtering query by %s: %s", filter_col, scope_name(o))
-            return query.filter(getattr(model, filter_col) == o.id)
-
-    logger.warn("query not filtered by scope")
-    return query
-
 
 
 EphemeralField = namedtuple('EphemeralField', "stringId isRelFld operStart startValue isNot isDisplay sortType formatName")
@@ -586,8 +498,20 @@ def build_query(session, collection, user, tableid, field_specs,
     query = QueryConstruct(
         collection=collection,
         objectformatter=ObjectFormatter(collection, user, replace_nulls),
-        query=session.query().distinct() if distinct else session.query(id_field),
+        query=session.query(),
     )
+
+    query, scope_fields = filter_by_collection(model, query, collection)
+
+    add_id = ((not distinct) or
+              can_uniquely_identify(field_specs,
+                                    datamodel.get_table_by_id(tableid, True).django_name,
+                                    [scope_fields]))
+    if add_id:
+        query = query.add_columns(id_field)
+
+    if distinct:
+        query = query.distinct()
 
     tables_to_read = set([
         table
@@ -598,7 +522,6 @@ def build_query(session, collection, user, tableid, field_specs,
     for table in tables_to_read:
         check_table_permissions(collection, user, table, "read")
 
-    query = filter_by_collection(model, query, collection)
 
     if recordsetid is not None:
         logger.debug("joining query to recordset: %s", recordsetid)
