+++ specifyweb/stored_queries/query_construct.py
@@ -109,5 +109,5 @@ def add_proxy_method(name):
         return self._replace(query=method(*args, **kwargs))
     setattr(QueryConstruct, name, proxy)
 
-for name in 'filter join outerjoin add_columns reset_joinpoint'.split():
+for name in 'filter join outerjoin add_columns reset_joinpoint distinct'.split():
     add_proxy_method(name)
